Re-organization and added notes/ideas file
