+++ project_root/notes/plans_text_files/data_preprocessing_plans.txt
@@ -1,0 +1,43 @@
+
+Data preprocessing structure plan/ideas
+
+1. Data Cleaning:
+Remove or interpolate missing or corrupted data points.
+Handle any anomalies or outliers that may affect the quality of your data.
+
+2. Noise Reduction:
+Apply filters to reduce noise and interference in the data.
+Use techniques like signal averaging or smoothing to reduce high-frequency noise.
+
+3. Normalization and Scaling:
+Normalize the data to a standard range, such as between 0 and 1 or -1 and 1.
+Scaling ensures that features have similar ranges, which can improve the training of neural networks.
+
+4. Resampling:
+Adjust the sampling rate if necessary to ensure consistency across data samples.
+Upsample or downsample the data to match the desired frequency or time resolution.
+
+5. Feature Extraction:
+Depending on your task, extract relevant features from the time-domain data. For RF spectrum data, consider the Fourier transform to analyze frequency domain characteristics.
+
+6. Data Splitting:
+Split your data into training, validation, and test sets. This is crucial for model evaluation.
+Implement stratified splitting if you have class-imbalanced data.
+
+7. Data Augmentation (Optional):
+Create augmented data samples by adding noise, perturbing the data, or introducing variations. This can increase the diversity of your training data.
+
+8. Label Encoding (For Classification):
+If your project involves classification tasks, encode categorical labels into numerical values.
+
+9. Handling Time Dependencies (Optional):
+If there are temporal dependencies in your data, consider incorporating previous time steps as input features or use recurrent neural networks (RNNs) or Long Short-Term Memory (LSTM) networks.
+
+10. Data Serialization (Optional):
+Serialize your preprocessed data into a format suitable for training, such as PyTorch tensors or NumPy arrays.
+
+11. Data Visualization (Optional):
+Create visualizations to explore and understand your data. This can be helpful for identifying patterns and anomalies.
+
+12. Saving Preprocessed Data:
+Save the preprocessed data to a file or data structure that you can easily load for model training.
