+++ project_root/training_script.py
@@ -1,0 +1,54 @@
+# Import statements
+import torch
+import torch.nn as nn
+import torch.optim as optim
+import torchvision.transforms as transforms  # For data preprocessing
+from torch.utils.data import DataLoader, TensorDataset
+from model_definition.algorithm_model import config, Algorithm_v0_1 # Import hyperparameter values
+
+
+
+# Extract hyperparameters
+input_size = config["input_size"]
+hidden_size = config["hidden_size"]
+output_size = config["output_size"]
+num_hidden_layers = config["num_hidden_layers"]
+learning_rate = config["learning_rate"]
+num_epochs = config["num_epochs"]
+batch_size = config["batch_size"]
+
+# Create instance of algorithm model
+model = Algorithm_v0_1(input_size, hidden_size, output_size, num_hidden_layers)
+
+# Apply the custom weight initialization function to the model
+model.init_weights()
+
+# Load and preprocess your RF spectrum data
+
+# Define loss function and optimizer ### EDIT LATER
+criterion = nn.MSELoss()  # Example loss function for regression
+optimizer = optim.Adam(model.parameters(), lr=0.001)
+
+# Training
+# Variable initializations
+num_epochs = 50  # Adjust
+
+# Dataset initialization
+train_inputs = 0
+train_labels = 0
+
+# Assuming you have your data as tensors
+train_data = TensorDataset(train_inputs, train_labels)
+batch_size = 32  # Adjust the batch size as needed
+
+# Training loop
+train_loader = DataLoader(train_data, batch_size=batch_size, shuffle=True)
+for epoch in range(num_epochs):
+    for data in train_loader:  # Iterate through training data
+        inputs, labels = data
+        optimizer.zero_grad()
+        outputs = model(inputs)
+        loss = criterion(outputs, labels)
+        loss.backward()
+        optimizer.step()
+    # Validation and monitoring during training
